Refactor: Animation Sequence Viewer 코드 구조, 가독성 개선
Changes to be committed:
	modified:   JSEngine/JSEngine.vcxproj
	modified:   JSEngine/JSEngine.vcxproj.filters
	modified:   JSEngine/Source/Editor/Animation/AnimationSequenceEditorDocument.cpp
	modified:   JSEngine/Source/Editor/Animation/AnimationSequenceEditorDocument.h
	modified:   JSEngine/Source/Editor/Animation/AnimationSequenceEditorWidget.cpp
	modified:   JSEngine/Source/Editor/Animation/AnimationSequenceEditorWidget.h
	new file:   JSEngine/Source/Editor/Animation/AnimationSequenceViewerBottomStrip.cpp
	new file:   JSEngine/Source/Editor/Animation/AnimationSequenceViewerPreviewPanel.cpp
	new file:   JSEngine/Source/Editor/Animation/AnimationSequenceViewerSequencerPanel.cpp
	new file:   JSEngine/Source/Editor/Animation/AnimationSequenceViewerUiHelpers.cpp
	new file:   JSEngine/Source/Editor/Animation/AnimationSequenceViewerUiHelpers.h
	modified:   JSEngine/Source/Engine/Animation/AnimInstance.cpp
	modified:   JSEngine/Source/Engine/Render/Mesh/VertexFactory/SkeletalVertexFactory.cpp

+++ JSEngine/Source/Engine/Animation/AnimInstance.cpp
@@ -703,7 +703,7 @@ FQuat UAnimInstance::InterpolateKeys(const TArray<FQuat>& Keys, float Time, floa
 void UAnimInstance::BlendPoses(const FSkeletonPose& PoseA, const FSkeletonPose& PoseB, float BlendFactor, FSkeletonPose& OutPose)
 {
     SCOPE_STAT_ANIM("Pose Blend");
-    const int32 BoneCount = PoseA.LocalTransforms.size();
+    const int32 BoneCount = static_cast<int32>(PoseA.LocalTransforms.size());
     OutPose.LocalTransforms.resize(BoneCount);
 
     for (int32 i = 0; i < BoneCount; ++i)
